feat: `u8` syscall ids

+++ eth-riscv-runtime/src/lib.rs
@@ -41,7 +41,7 @@ use eth_riscv_syscalls::Syscall;
 
 pub fn return_riscv(addr: u64, offset: u64) -> ! {
     unsafe {
-        asm!("ecall", in("a0") addr, in("a1") offset, in("t0") u32::from(Syscall::Return));
+        asm!("ecall", in("a0") addr, in("a1") offset, in("t0") u8::from(Syscall::Return));
     }
     unreachable!()
 }
@@ -49,26 +49,26 @@ pub fn return_riscv(addr: u64, offset: u64) -> ! {
 pub fn sload(key: u64) -> u64 {
     let value: u64;
     unsafe {
-        asm!("ecall", lateout("a0") value, in("a0") key, in("t0") u32::from(Syscall::SLoad));
+        asm!("ecall", lateout("a0") value, in("a0") key, in("t0") u8::from(Syscall::SLoad));
     }
     value
 }
 
 pub fn sstore(key: u64, value: u64) {
     unsafe {
-        asm!("ecall", in("a0") key, in("a1") value, in("t0") u32::from(Syscall::SStore));
+        asm!("ecall", in("a0") key, in("a1") value, in("t0") u8::from(Syscall::SStore));
     }
 }
 
 pub fn call(addr: u64, value: u64, in_mem: u64, in_size: u64, out_mem: u64, out_size: u64) {
     unsafe {
-        asm!("ecall", in("a0") addr, in("a1") value, in("a2") in_mem, in("a3") in_size, in("a4") out_mem, in("a5") out_size, in("t0") u32::from(Syscall::Call));
+        asm!("ecall", in("a0") addr, in("a1") value, in("a2") in_mem, in("a3") in_size, in("a4") out_mem, in("a5") out_size, in("t0") u8::from(Syscall::Call));
     }
 }
 
 pub fn revert() -> ! {
     unsafe {
-        asm!("ecall", in("t0") u32::from(Syscall::Revert));
+        asm!("ecall", in("t0") u8::from(Syscall::Revert));
     }
     unreachable!()
 }
@@ -88,7 +88,7 @@ pub fn keccak256(offset: u64, size: u64) -> B256 {
             lateout("a1") second,
             lateout("a2") third,
             lateout("a3") fourth,
-            in("t0") u32::from(Syscall::Keccak256)
+            in("t0") u8::from(Syscall::Keccak256)
         );
     }
 
@@ -107,7 +107,7 @@ pub fn msg_sender() -> Address {
     let second: u64;
     let third: u64;
     unsafe {
-        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, in("t0") u32::from(Syscall::Caller));
+        asm!("ecall", lateout("a0") first, lateout("a1") second, lateout("a2") third, in("t0") u8::from(Syscall::Caller));
     }
     let mut bytes = [0u8; 20];
     bytes[0..8].copy_from_slice(&first.to_be_bytes());
