+++ eth-riscv-syscalls/src/lib.rs
@@ -3,7 +3,7 @@
 macro_rules! syscalls {
     ($(($num:expr, $identifier:ident, $name:expr)),* $(,)?) => {
         #[derive(Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
-        #[repr(u32)]
+        #[repr(u8)]
         pub enum Syscall {
             $($identifier = $num),*
         }
@@ -26,15 +26,15 @@ macro_rules! syscalls {
             }
         }
 
-        impl From<Syscall> for u32 {
+        impl From<Syscall> for u8 {
             fn from(syscall: Syscall) -> Self {
                 syscall as Self
             }
         }
 
-        impl core::convert::TryFrom<u32> for Syscall {
+        impl core::convert::TryFrom<u8> for Syscall {
             type Error = ();
-            fn try_from(value: u32) -> Result<Self, Self::Error> {
+            fn try_from(value: u8) -> Result<Self, Self::Error> {
                 match value {
                     $($num => Ok(Syscall::$identifier)),*,
                     _ => Err(()),
